+++ CHANGES.rst
@@ -10,6 +10,9 @@ Changelog
 - Added basic tests to package.
   [loechel]
 
+- Changed parser.py to use functions from zc.buildout to get versions and versionannotations section names.
+  [loechel]
+
 - Add function to extract date information from PyPI to analyze package age.
   [loechel]